requirements.txt
Flask>=2.0
Pillow>=9.0
python-dotenv>=0.15
gunicorn>=20.0
huggingface_hub
requests
urllib3
google-cloud-secret-manager